+++ L3 Automation/python_guis/rip/rip_network_add_gui.py
@@ -3,7 +3,7 @@ import tkinter as tk
 from tkinter import messagebox
 
 from resources.devices.Router import Router
-from gui_resources import config
+from python_guis.gui_resources import config
 from resources.routing_protocols.Network import Network
